requirements.txt
streamlit
langchain
openai
unstructured[all-docs]
python-docx 
python-pptx
PyPDF2